Asbtracts out calculateOverviewData and adds tests.
