+++ assets/js/modules/analytics/util/index.js
@@ -29,7 +29,10 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { changeToPercent, getModulesData } from '../../../util';
+import { getModulesData } from '../../../util';
+import calculateOverviewData from './calculateOverviewData';
+
+export { calculateOverviewData };
 
 export const extractAnalyticsDataForTrafficChart = ( reports ) => {
 	if ( ! reports || ! reports.length ) {
@@ -208,44 +211,6 @@ export const extractAnalyticsDashboardSparklineData = ( reports ) => {
 	} );
 
 	return dataMap;
-};
-
-export const calculateOverviewData = ( reports ) => {
-	if ( ! reports || ! reports.length ) {
-		return false;
-	}
-
-	const { totals } = reports[ 0 ].data;
-	const lastMonth = totals[ 0 ].values;
-	const previousMonth = totals[ 1 ].values;
-
-	const totalUsers = lastMonth[ 0 ];
-	const totalSessions = lastMonth[ 1 ];
-	const averageBounceRate = lastMonth[ 2 ];
-	const averageSessionDuration = lastMonth[ 3 ];
-	const goalCompletions = lastMonth[ 4 ];
-	const totalPageViews = lastMonth[ 5 ];
-	const totalUsersChange = changeToPercent( previousMonth[ 0 ], lastMonth[ 0 ] );
-	const totalSessionsChange = changeToPercent( previousMonth[ 1 ], lastMonth[ 1 ] );
-	const averageBounceRateChange = changeToPercent( previousMonth[ 2 ], lastMonth[ 2 ] );
-	const averageSessionDurationChange = changeToPercent( previousMonth[ 3 ], lastMonth[ 3 ] );
-	const goalCompletionsChange = changeToPercent( previousMonth[ 4 ], lastMonth[ 4 ] );
-	const totalPageViewsChange = changeToPercent( previousMonth[ 5 ], lastMonth[ 5 ] );
-
-	return {
-		totalUsers,
-		totalSessions,
-		averageBounceRate,
-		averageSessionDuration,
-		totalUsersChange,
-		totalSessionsChange,
-		averageBounceRateChange,
-		averageSessionDurationChange,
-		goalCompletions,
-		goalCompletionsChange,
-		totalPageViews,
-		totalPageViewsChange,
-	};
 };
 
 /**
